fixed some compiler warnings with printf() format strings and others misc warnings
git-svn-id: https://svn.aros.org/svn/aros/trunk@8781 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/demos/filereq.c
@@ -147,7 +147,7 @@ static void showrequester(char *msg, struct TagItem *tags)
 
 		for(i = 1; i <= fr->fr_NumArgs; i++)
 		{
-		    printf("%3ld: %s\n", i, wbarg->wa_Name);
+		    printf("%3ld: %s\n", (long)i, wbarg->wa_Name);
 		    wbarg++;
 		}
 	    }
